sti-docs/adaptive-learning-service.py
import tensorflow as tf
import numpy as np
from typing import Dict, Any

class AdaptiveLearningService:
    def __init__(self, student_profile: Dict[str, Any]):
        self.student_profile = student_profile
        self.model = self._build_adaptive_model()
    
    def _build_adaptive_model(self):
        """
        Cria um modelo de rede neural para personalização de aprendizado
        """
        model = tf.keras.Sequential([
            tf.keras.layers.Dense(64, activation='relu', input_shape=(len(self.student_profile),)),
            tf.keras.layers.Dense(32, activation='relu'),
            tf.keras.layers.Dense(16, activation='relu'),
            tf.keras.layers.Dense(1, activation='sigmoid')
        ])
        
        model.compile(optimizer='adam', loss='binary_crossentropy')
        return model
    
    def personalize_content(self, available_content):
        """
        Personaliza conteúdo baseado no perfil do estudante
        """
        difficulty_scores = self._calculate_difficulty_scores(available_content)
        recommended_content = self._recommend_content(difficulty_scores)
        
        return recommended_content
    
    def _calculate_difficulty_scores(self, content):
        # Lógica para calcular scores de dificuldade
        pass
    
    def _recommend_content(self, scores):
        # Lógica para recomendar conteúdo
        pass
